Move asmop and asmregister to asm module.
