+++ sway-core/src/ir_generation/function.rs
@@ -13,7 +13,7 @@ use crate::{
     ir_generation::const_eval::{
         compile_constant_expression, compile_constant_expression_to_constant,
     },
-    language::{parsed::*, *},
+    language::*,
     metadata::MetadataManager,
     semantic_analysis::*,
     type_system::{look_up_type_id, to_typeinfo, IntegerBits, TypeId, TypeInfo},
