+++ sway-core/src/semantic_analysis/ast_node/expression/typed_expression_variant.rs
@@ -1,8 +1,4 @@
-use crate::{
-    language::{parsed::*, *},
-    semantic_analysis::*,
-    type_system::*,
-};
+use crate::{language::*, semantic_analysis::*, type_system::*};
 
 use sway_types::{state::StateIndex, Ident, Span, Spanned};
